Reorganize instruction variants
This makes the binary and text variants synchronized. In particular,
they use matching types with matching names for all immediates. The only
exception to this rule is `Index` vs. `text::Var`.

Changes:

* Use the name `SimdLaneImmediate` instead of `u8`
* Use the name `SelectImmediate` instead of `ValueTypeList`
* Rename `has_u8_immediate` -> `has_simd_lane_immediate`
* Rename `has_value_type_list_immediate` -> `has_select_immediate`
* Don't store `u8` immediate for reserved values (e.g. `memory.grow`)
* Use `s32` and `s64` for the `i32.const` and `i64.const` immediates

+++ include/wasp/text/read.h
@@ -124,7 +124,7 @@ auto ReadNameEqNatOpt(Tokenizer&, Context&, TokenType, u32) -> OptAt<u32>;
 auto ReadAlignOpt(Tokenizer&, Context&) -> OptAt<u32>;
 auto ReadOffsetOpt(Tokenizer&, Context&) -> OptAt<u32>;
 
-auto ReadSimdLane(Tokenizer&, Context&) -> At<u32>;
+auto ReadSimdLane(Tokenizer&, Context&) -> At<u8>;
 auto ReadSimdShuffleImmediate(Tokenizer&, Context&) -> At<ShuffleImmediate>;
 template <typename T, size_t N>
 auto ReadSimdValues(Tokenizer&, Context&) -> At<v128>;
